remove checkout flow flag
