+++ packages/client/components/Dashboard/MobileDashSidebar.tsx
@@ -95,7 +95,6 @@ const MobileDashSidebar = (props: Props) => {
       fragment MobileDashSidebar_viewer on User {
         ...StandardHub_viewer
         featureFlags {
-          checkoutFlow
           retrosInDisguise
         }
         organizations {
@@ -110,9 +109,8 @@ const MobileDashSidebar = (props: Props) => {
   )
   if (!viewer) return null
   const {featureFlags, organizations} = viewer
-  const showOrgSidebar = featureFlags.checkoutFlow && match
 
-  if (showOrgSidebar) {
+  if (match) {
     const {orgId: orgIdFromParams} = match.params
     const currentOrg = organizations.find((org) => org.id === orgIdFromParams)
     const {id: orgId, name, isBillingLeader} = currentOrg ?? {}
